Update rbt_device_rssi and rbt_device_link_quality
In OS X 10.7 and newer the old way of retrieving the RSSI and link
quality are no longer supported.  It appears that the extension's RSSI
method maps to [IOBluetoothDevice rawRSSI] and the link_quality method
to [IOBluetoothDevice RSSI].

The bluetooth extension no longer works on OS X 10.6 and earlier because
I don't know how to detect OS X methods in the framework.  When someone
complains the old behavior can be restored.